+++ src/ACF/Field_Group.php
@@ -3,7 +3,11 @@
 
 namespace Tribe\Libs\ACF;
 
-class Field_Group extends Sub_Field implements ACF_Aggregate {
+use Tribe\Libs\ACF\Traits\With_Sub_Fields;
+
+class Field_Group extends Field implements ACF_Aggregate {
+	use With_Sub_Fields;
+	
 	protected $key_prefix = 'field';
 
 	/**
@@ -13,5 +17,12 @@ class Field_Group extends Sub_Field implements ACF_Aggregate {
 	public function __construct( $key, $attributes = [] ) {
 		parent::__construct( $key, $attributes );
 		$this->attributes[ 'type' ] = 'group';
+	}
+
+	public function get_attributes() {
+		$attributes                 = parent::get_attributes();
+		$attributes[ 'sub_fields' ] = $this->get_sub_field_attributes();
+
+		return $attributes;
 	}
 }
